Split out mergePostEdits
